10/4 15:58 ComunicacionEmisor, ComunicacionReceptor
